+++ src/main/java/co/istad/dealkh/features/wishlist/dto/WishListResponse.java
@@ -1,11 +1,13 @@
 package co.istad.dealkh.features.wishlist.dto;
 
+import java.math.BigDecimal;
+
 public record WishListResponse(
         String uuid,
         String profile,
         String productName,
         String username,
-        Double discountPercentage,
+        BigDecimal discountPercentage,
         String description,
         String isGranted
 ) {
